tests/wpt/tests/css/cssom/caretRangeFromPoint-textarea-transform.tentative.html

<!doctype html>
<meta charset="utf-8" />
<title>document.caretRangeFromPoint()</title>
<link rel="help" href="https://github.com/w3c/csswg-drafts/pull/12362" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  textarea {
    display: block;
    translate: -50%;
  }
</style>
<textarea id="textarea">
AAAAAAAAAAAAAAAAAAAAA</textarea
>
<script>
  test(() => {
    let range = document.caretRangeFromPoint(0, 0);
    assert_true(range instanceof Range);
    assert_equals(range.startOffset, 0);
    assert_equals(range.endOffset, 0);
    assert_equals(range.startContainer, textarea);
    assert_equals(range.endContainer, textarea);
  }, "document.caretRangeFromPoint(0, 0)");
  test(() => {
    let range = document.caretRangeFromPoint(10, 10);
    assert_true(range instanceof Range);
    assert_equals(range.startOffset, 0);
    assert_equals(range.endOffset, 0);
    assert_equals(range.startContainer, textarea);
    assert_equals(range.endContainer, textarea);
  }, "document.caretRangeFromPoint(10, 10)");
</script>
