LayoutTests/imported/w3c/web-platform-tests/css/cssom/caretRangeFromPoint-textarea-transform.tentative-expected.txt



FAIL document.caretRangeFromPoint(0, 0) assert_equals: expected Element node <textarea id="textarea">AAAAAAAAAAAAAAAAAAAAA</textarea> but got Element node <body><textarea id="textarea">AAAAAAAAAAAAAAAAAAAAA</text...
PASS document.caretRangeFromPoint(10, 10)
